Rename zshrc configs
Using a dot is so much more pleasing than a hyphen.

I use Linux more these days so I should relegate the macos config to
also having a suffix.

This adds goenv config.

I've disabled the GOPATH management from goenv since it prefixes the
GOPATH with the version of Go. I haven't quite worked out how what this
means for working with a project repository that has packages inside it.